ci(quality): split planner invariants module; re-record function ceilings
planner/mod.rs crossed the 750-line module cap when the
anchor_element_id wrapper + audit row landed — the post-pass
structural invariant checkers (debug_check_invariants + three
check_* fns, ~100 lines, a self-contained concern) move to
planner/invariants.rs rather than raising the cap.

Function-ceiling refresh: 2 tightened, 2 raised — the raises are the
scalar dispatchers' +14/+18 branches from the zero-arg first_arg
guards (one ? per previously-panicking site), the recorded shape of
the panic fix, not drift.
